app/requirements.txt
alembic==1.7.5
fastapi==0.85.1
uvicorn==0.16.0
pydantic==1.10.4
requests==2.26.0
python-dotenv==0.21.0
# psycopg2==2.9.9
async-timeout==4.0.3
asyncpg==0.29.0
pytz==2024.1
email-validator==2.2.0
bcrypt==4.2.0
PyJWT==2.9.0
aiosmtplib==3.0.2